Update PermutationRecursion.java
Hello Dana! I did a little bit change to this. There seems no need to return a String for permutateHelper. If just changed it to void, there will be duplicate outputs. So I changed it to an if-else structure, so that it can also output correctly. Hope you like it : )